+++ lib/origami-build-tools-cli.js
@@ -19,12 +19,12 @@ const help = `
 			$ obt <command> [<options>]
 
 		Commands
-			build, b          Build CSS and JS in current directory
-			boilerplate, bp   Build a boilerplate folder structure for a new component
-			demo, d           Build demos into the demos directory
-			install, i        Install npm and bower dependencies required to build modules
-			test, t           Run Origami specification tests and component specific tests
-			verify, v         Check folder and code structure follows Origami specification
+			build, b       Build CSS and JS in current directory
+			demo, d        Build demos into the demos directory
+			init           Initialise a new component with a boilerplate folder structure
+			install, i     Install npm and bower dependencies required to build modules
+			test, t        Run Origami specification tests and component specific tests
+			verify, v      Check folder and code structure follows Origami specification
 
 		Options
 			-h, --help                 Print out this message
@@ -67,9 +67,8 @@ switch (argument) {
 	case 'b':
 		task = require('./tasks/build');
 		break;
-	case 'boilerplate':
-	case 'bp':
-		task = require('./tasks/boilerplate');
+	case 'init':
+		task = require('./tasks/init');
 		break;
 	case 'demo':
 	case 'd':
