+++ lib/tasks/boilerplate.js
@@ -1,38 +1,44 @@
 'use strict';
 
-const Listr = require('listr');
+const fs = require('fs-extra');
+const path = require('path');
+const template = require('./../../templates/boilerplate.js');
 
-const buildBoilerplateTree = require('./boilerplate-tree');
+const buildTree = (root, name) => {
+	const generate = (file, template) => fs.outputFile(path.join(root, file), template);
 
-function camelCase (string) {
-	return string.replace(/\-+(.)/g, (match, chr) => { return chr.toUpperCase(); });
-}
-
-function titleCase (string) {
-	return string.toLowerCase().split('-').map(function(word) {
-		return word.replace(word[0], word[0].toUpperCase());
-	}).join('');
-}
-
-module.exports = function (cfg) {
-	cfg = cfg || {};
-	const config = cfg.flags || {};
-	config.cwd = config.cwd || process.cwd();
-	config.name = cfg.input && cfg.input.length > 1 ? cfg.input[1] : 'o-component-boilerplate';
-	config.name = {
-		original: config.name,
-		camelCase: camelCase(config.name),
-		titleCase: titleCase(config.name)
-	};
-
-	return new Listr([{
-		title: `Building '${config.name.original}' boilerplate`,
-		task: () => buildBoilerplateTree(config),
-		skip: () => false
-	}], {
-		renderer: require('../helpers/listr-renderer')
-	})
-		.run();
+	return Promise.all([
+		generate('/.gitignore', template.gitIgnore()),
+		generate('/main.js', template.mainJs(name)),
+		generate('/main.scss', template.mainScss(name)),
+		generate('/origami.json', template.origamiJson(name)),
+		generate('/README.md', template.readMe(name)),
+		generate('/.circleci/config.yml', template.ciConfigYml()),
+		generate(`/src/js/${name.original}.js`, template.srcJs(name)),
+		generate('/src/scss/_functions.scss', ''),
+		generate('/src/scss/_mixins.scss', template.srcSassMixins(name)),
+		generate('/src/scss/_variables.scss', template.srcSassVariables(name)),
+		generate('/demos/src/demo.js', template.demoJs()),
+		generate('/demos/src/demo.mustache', template.demoMustache(name)),
+		generate('/demos/src/pa11y.mustache', template.demoPa11y(name)),
+		generate('/demos/src/demo.scss', template.demoSass(name)),
+		generate(`/test/${name.camelCase}.test.js`, template.testMain(name)),
+		generate(`/test/helpers/fixtures.js`, template.testFixtures(name)),
+		generate(`/test/scss/index.test.scss`, template.testSassMain(name)),
+		generate(`/test/scss/_mixins.test.scss`, template.testSassMixins(name))
+	]);
 };
 
-module.exports.watchable = false;
+module.exports = (config) => {
+	const name = config.name;
+	const root = path.join(config.cwd, name.original);
+
+	return fs.pathExists(root)
+		.then(exists => {
+			if (exists) {
+				return Promise.reject(new Error(`'${name.original}' already exists in this directory.`));
+			} else {
+				buildTree(root, name);
+			}
+		});
+};
