+++ lib/tasks/init.js
@@ -1,0 +1,38 @@
+'use strict';
+
+const Listr = require('listr');
+
+const buildBoilerplate = require('./boilerplate');
+
+function camelCase (string) {
+	return string.replace(/\-+(.)/g, (match, chr) => { return chr.toUpperCase(); });
+}
+
+function titleCase (string) {
+	return string.toLowerCase().split('-').map(function(word) {
+		return word.replace(word[0], word[0].toUpperCase());
+	}).join('');
+}
+
+module.exports = function (cfg) {
+	cfg = cfg || {};
+	const config = cfg.flags || {};
+	config.cwd = config.cwd || process.cwd();
+	config.name = cfg.input && cfg.input.length > 1 ? cfg.input[1] : 'o-component-boilerplate';
+	config.name = {
+		original: config.name,
+		camelCase: camelCase(config.name),
+		titleCase: titleCase(config.name)
+	};
+
+	return new Listr([{
+		title: `Building '${config.name.original}' boilerplate`,
+		task: () => buildBoilerplate(config),
+		skip: () => false
+	}], {
+		renderer: require('../helpers/listr-renderer')
+	})
+		.run();
+};
+
+module.exports.watchable = false;
